xProject/requirements.txt
click==6.7
Flask==0.12.2
Flask-Cors==3.0.3
Flask-SQLAlchemy==2.3.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
paho-mqtt==1.3.1
psycopg2==2.7.3.2
six==1.11.0
SQLAlchemy==1.2.0
Werkzeug==0.14.1
